requirements.txt
pydantic
pyzmq
